# rks-openapi.yaml
...
openapi: 3.0.0
info:
  title: Remote Key Server API
  description: |
    The _Remote Key Server_ is a solution to manage TLS private keys and certificates in a distributed system.

    Distribution of sensitive material, especially private keys, requires to take care of the access, the storage, and the duration of the storage of this secrets on server nodes.

    In the context of a Content Delivery Network, HTTPS traffic requires to handle multiple content providers certificates and private_keys and distribute them accross all the CDN nodes. This leads to difficult challenges over the secret managements and particularly security concerns with nodes storing every secret locally for an indefinite amount of time.

    The RKS proposes a simple model to
      * secure secret distribution by limiting secret access to registered nodes only
      * limit secret storage duration on a node to a configurable time to live
      * ease secret management by providing a single configuration endpoint via API
      * add crisis management solution to cut access to the RKS to specific nodes or an entire CDN in case of emergency

    The RKS is based on the Hashicorp Vault Open Source Software (https://github.com/hashicorp/vault) and provides an API on top of Vault's
  version: 0.9.0
...
security:
  - X-Vault-Token: []
paths:
...
  '/rks/v1/group/{groupname}':
    post:
      summary: Create a Group of Nodes
      description: |
        Authenticate with **admin_token**
      tags:
        - RKS Administration
      operationId: createGroup
      parameters:
        - $ref: '#/components/parameters/groupname'
      requestBody:
        description: Group registration information
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GroupRegInfo'
      responses:
        '201':
          description: Group successfully registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GroupToken'
        '403':
          description: "Forbidden. Invalid X-Vault-Token"
        '409':
          description: Conflict Group already exists
    get:
      summary: Get registration information for a Group of Nodes
      description: |
        Authenticate with **admin_token**
      tags:
        - RKS Administration
      operationId: getGroup
      parameters:
        - $ref: '#/components/parameters/groupname'
      responses:
        '200':
          description: Group registration information successfully retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GroupRegInfo'
        '403':
          description: "Forbidden. Invalid X-Vault-Token"
        '404':
          description: "Group not found"
    put:
      summary: Update a Group of Nodes
      description: |
        Authenticate with **admin_token**
      tags:
        - RKS Administration
      operationId: updateGroup
      parameters:
        - $ref: '#/components/parameters/groupname'
      requestBody:
        description: Group registration information
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GroupRegInfo'
      responses:
        '200':
          description: Group registration information successfully updated
        '403':
          description: "Forbidden. Invalid X-Vault-Token"
        '404':
          description: "Group not found"
    delete:
      summary: Delete a Group of Nodes
      description: |
        Authenticate with **admin_token**
      tags:
        - RKS Administration
      operationId: deleteGroup
      parameters:
        - $ref: '#/components/parameters/groupname'
      responses:
        '204':
          description: Group successfully deleted
        '403':
          description: "Forbidden. Invalid X-Vault-Token"
        '404':
          description: Groupname Not Found
  '/rks/v1/group/{groupname}/grouptoken':
    get:
      summary: Get grouptoken for this **groupname**
      description: |
        Authenticate with **admin_token** retrieved through previous login
        request
      tags:
        - RKS Administration
      operationId: GetGroupToken
      parameters:
        - $ref: '#/components/parameters/groupname'
      responses:
        '200':
          description: grouptoken successfully retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GroupToken'
        '403':
          description: "Forbidden. Invalid X-Vault-Token"
        '404':
          description: Group not found
    put:
      summary: Update grouptoken for this **groupname**
      description: |
        Authenticate with **admin_token** retrieved through previous login
        request
      tags:
        - RKS Administration
      operationId: UpdateGroupToken
      parameters:
        - $ref: '#/components/parameters/groupname'
      responses:
        '200':
          description: grouptoken successfully updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GroupToken'
        '403':
          description: "Forbidden. Invalid X-Vault-Token"
        '404':
          description: Group not found
  '/rks/v1/group/{groupname}/config':
    get:
      summary: Get config for this **groupname**
      description: |
        Authenticate with **admin_token** retrieved through previous login
        request
      tags:
        - RKS Administration
      operationId: GetGroupConfig
      parameters:
        - $ref: '#/components/parameters/groupname'
      responses:
        '200':
          description: config successfully retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GroupRegInfo'
        '403':
          description: "Forbidden. Invalid X-Vault-Token"
        '404':
          description: Group not found
  '/rks/v1/group/{groupname}/nodes/{nodeId}':
    delete:
      summary: Revoke a Node by revoking its token
      description: revoke token for node of **groupname** with nodeId like **X-LCDN-nodeId**,  
        **adminToken** is used as X-Vault-Token for authentication. Note that this call will
       only revoke Node Token, that's to say, if you want to disallow this node to register again,
       there is two ways. 
       Makes callback url answer "40X" for this nodeId
       Renew grouptoken and spread it to all nodes, execept this one 
      tags:
        - RKS Administration
      operationId: revokeNode
      parameters:
        - $ref: '#/components/parameters/groupname'
        - $ref: '#/components/parameters/nodeIdPath'
      responses:
        '204':
          description: Node successfully deleted
        '403':
          description: "Forbidden. Invalid X-Vault-Token"
        '404':
          description: Group not found
  '/rks/v1/group/{groupname}/secrets':
    get:
      summary: Get secrets associated with **groupname**
      description: |
        Authenticate with **admin_token** retrieved through previous login
        request
      tags:
        - RKS Administration
      operationId: GetGroupSecrets
      parameters:
        - $ref: '#/components/parameters/groupname'
      responses:
        '200':
          description: Ok
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
                  example: "test.com"
        '403':
          description: "Forbidden. Invalid X-Vault-Token"
  '/rks/v1/group/{groupname}/secrets/{fqdn}':
    post:
      summary: Associate **fqdn** secret with **groupname**
      description: |
        Authenticate with **admin_token** retrieved through previous login
        request
      tags:
        - RKS Administration
      operationId: AssociateSecret
      parameters:
        - $ref: '#/components/parameters/groupname'
        - $ref: '#/components/parameters/fqdn'
      responses:
        '201':
          description: FQDN Associated
        '403':
          description: "Forbidden. Invalid X-Vault-Token"
        '404':
          description: Secret or groupName not Found
    delete:
      summary: Dissociate **fqdn** secret from **groupname**
      description: |
        Authenticate with **admin_token** retrieved through previous login
        request
      tags:
        - RKS Administration
      operationId: DissociateSecret
      parameters:
        - $ref: '#/components/parameters/groupname'
        - $ref: '#/components/parameters/fqdn'
      responses:
        '204':
          description: FQDN un-associated
        '403':
          description: "Forbidden. Invalid X-Vault-Token"
        '404':
          description: Secret or group not found
...
  '/rks/v1/secret/{fqdn}':
    get:
      summary: Get fqdn secret
      description: |
        Get Secret for **fqdn**\
        **nodeToken** or **adminToken** is used as X-Vault-Token for authentication
      operationId: getSecret
      tags:
        - RKS Administration
        - RKS Secrets
      parameters:
        - $ref: '#/components/parameters/fqdn'
      responses:
        '200':
          description: Ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Secret'
        '403':
          description: |
            Forbidden. Invalid X-Vault-Token
        '404':
          description: Not Found
    post:
      summary: Create fqdn secret
      description: |
        Create Secret for **fqdn**\
        **adminToken** is used as X-Vault-Token for authentication
      operationId: createSecret
      tags:
        - RKS Administration
      parameters:
        - $ref: '#/components/parameters/fqdn'
      requestBody:
        description: Secret
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Secret'
      responses:
        '204':
          description: Secret created
        '403':
          description: |
            Forbidden. Invalid X-Vault-Token
        '409':
          description: Conflict Secret already exists
    put:
      summary: Update fqdn secret
      description: |
        Update Secret for **fqdn**\
        **adminToken** is used as X-Vault-Token for authentication
      operationId: updateSecret
      tags:
        - RKS Administration
      parameters:
        - $ref: '#/components/parameters/fqdn'
      requestBody:
        description: Secret
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Secret'
      responses:
        '200':
          description: Secret updated
        '403':
          description: |
            "Forbidden. Invalid X-Vault-Token"
        '404':
          description: Not Found
    delete:
      summary: Delete fqdn secret
      description: |
        Delete Secret for **fqdn**\
        **adminToken** is used as X-Vault-Token for authentication
      operationId: deleteSecret
      tags:
        - RKS Administration
      parameters:
        - $ref: '#/components/parameters/fqdn'
      responses:
        '204':
          description: secret deleted
        '403':
          description: |
            Forbidden. **adminToken** does not allow to retrieve **fqdn**
        '404':
          description: Not Found
        '409':
          description: Conflict, Secret associated with at least one group
  '/rks/v1/secret/{fqdn}/groups':
    get:
      summary: Get list of groupname associated to this fqdn secret
      description: |
        Get list of groupname associated to Secret with **fqdn**\
        **adminToken** is used as X-Vault-Token for authentication
      operationId: getSecretGroups
      tags:
        - RKS Administration
      parameters:
        - $ref: '#/components/parameters/fqdn'
      responses:
        '200':
          description: Ok
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
                  example: "cdn1"
        '403':
          description: Forbidden. Invalid X-Vault-Token
        '404':
          description: secret Not Found
components:
  parameters:
    groupname:
      name: groupname
      in: path
      required: true
      description: GroupName
      schema:
        type: string
        example: "cdn1"
        pattern: "^[a-zA-Z0-9\\-]{1,64}$"
    fqdn:
      name: fqdn
      in: path
      required: true
      description: FQDN
      schema:
        type: string
        example: "test.com"
    nodeIdPath:
      name: nodeId
      in: path
      required: true
      description: |
        **nodeID**
      schema:
        type: string
        example: "8c03fb4a-007f-11ea-9383-0fd702e68c30"
        pattern: "^[a-zA-Z0-9\\-]{1,64}$"
...
  securitySchemes:
    X-Vault-Token:
      type: apiKey
      in: header
      name: X-Vault-Token
      description: |
        X-Vault-Token is a token obtained from Vault

        In the RKS context it can either be a **rootToken** (for initialization), an **adminToken**, a **groupToken** or a **nodeToken** depending on the API endpoint
  schemas:
...
    GroupToken:
      properties:
        groupToken:
          type: string
          example: "5Rgiycc79m1oNsNxCM3WAHe5"
...
    GroupRegInfo:
      required:  
        - callbackURL
        - oauthURL
        - oauthClientID
        - oauthClientSecret
      properties:
        callbackURL:
          type: string
          example: "https://group-manager.group:3000/prefix"
        oauthURL:
          type: string
          example: "https://group-manager.group:3000/api/identity/oauth/token"
        oauthClientID:
          type: string
          example: "f146c93f-00af-42a2-9667-8c0d99b37953"
        oauthClientSecret:
          type: string
          example: "da5edef7-1664-42b9-8fa7-5a107e7cca32"
...